+++ scripts/_settings.mjs
@@ -152,14 +152,6 @@ Hooks.once('setup', () => {
 		default: true,
 		config: true
 	}
-	const enableGlobalConflicts = {
-		name: `${MODULE.ID}.settings.enableGlobalConflicts.name`,
-		hint: `${MODULE.ID}.settings.enableGlobalConflicts.hint`,
-		type: Boolean,
-		default: false,
-		scope: 'world',
-		config: true
-	}
 	const disableCoreConflicts = {
 		name: `${MODULE.ID}.settings.disableCoreConflicts.name`,
 		hint: `${MODULE.ID}.settings.disableCoreConflicts.hint`,
@@ -168,13 +160,30 @@ Hooks.once('setup', () => {
 		scope: 'world',
 		config: false
 	}
+	const enableGlobalConflicts = {
+		name: `${MODULE.ID}.settings.enableGlobalConflicts.name`,
+		hint: `${MODULE.ID}.settings.enableGlobalConflicts.hint`,
+		type: Boolean,
+		default: false,
+		scope: 'world',
+		config: true
+	}
+	const addGoogleSheetButton = {
+		name: `${MODULE.ID}.settings.addGoogleSheetButton.name`,
+		hint: `${MODULE.ID}.settings.addGoogleSheetButton.hint`,
+		type: Boolean,
+		default: false,
+		scope: 'world',
+		config: true
+	}
 	if (game.modules.get('lib-server-setting')?.active ?? false) {
 		Hooks.once('lib-server-setting.Setup', async (SETTING) => {
 			SETTING(MODULE.ID, 'trackedChangelogs', trackedChangelogs);
 			SETTING(MODULE.ID, 'showNewChangelogsOnLoad', showNewChangelogsOnLoad);
 			SETTING(MODULE.ID, 'renamedModules', renamedModules);
-			SETTING(MODULE.ID, 'enableGlobalConflicts', enableGlobalConflicts);
 			SETTING(MODULE.ID, 'disableCoreConflicts', disableCoreConflicts);
+			SETTING(MODULE.ID, 'enableGlobalConflicts', enableGlobalConflicts);
+			SETTING(MODULE.ID, 'addGoogleSheetButton', addGoogleSheetButton);
 			SETTING(MODULE.ID, 'autoPrefixModules', autoPrefixModules);
 			SETTING(MODULE.ID, 'smartPrefix', smartPrefix);
 			SETTING(MODULE.ID, 'presets', presets);
@@ -183,8 +192,9 @@ Hooks.once('setup', () => {
 		MODULE.setting('register', 'trackedChangelogs', trackedChangelogs);
 		MODULE.setting('register', 'showNewChangelogsOnLoad', showNewChangelogsOnLoad);
 		MODULE.setting('register', 'renamedModules', renamedModules);
-		MODULE.setting('register', 'enableGlobalConflicts', enableGlobalConflicts);
 		MODULE.setting('register', 'disableCoreConflicts', disableCoreConflicts);
+		MODULE.setting('register', 'enableGlobalConflicts', enableGlobalConflicts);
+		MODULE.setting('register', 'addGoogleSheetButton', addGoogleSheetButton);
 		MODULE.setting('register', 'autoPrefixModules', autoPrefixModules);
 		MODULE.setting('register', 'smartPrefix', smartPrefix);
 		MODULE.setting('register', 'presets', presets);